+++ functions/eslint.config.js
@@ -18,6 +18,8 @@ module.exports = [
         process: "readonly",
         setTimeout: "readonly",
         clearTimeout: "readonly",
+        setInterval: "readonly",
+        clearInterval: "readonly",
         Buffer: "readonly",
         Promise: "readonly",
         URL: "readonly",